Fix 43 logic bugs in function, tester, stringOp, technology, structure, aspect and flow
Covers all 42 rows recorded for these folders in HANDOFF.md, plus one
undocumented sibling: FuzzyDictionary.getMostSimilarItem(Object) seeded
its running minimum at 0 and tested 'minDist > distance', so with
non-negative distances it always returned -1. Without that fix the
documented two-arg row would have stayed broken, since it delegates here.

Main defect families:
- ConditionTable's constructor validated the still-null instance field
  instead of its parameter, an NPE on every construction. LinAt, CatProcessor
  and structure/aspect/ListAspect had the same field-instead-of-parameter
  defect.
- Unimplemented stubs that silently returned a wrong answer rather than
  failing: SentenceComparer.getWordSet and getMostSimilarSentence,
  Ternary.fromString, Graph2D-style 'not implemented' returns.
- Off-by-one loop bounds: Grammar.evolve, ListAspect.removeVal,
  HistoryList.addItem, Network.randomizeWeights, Enum.succ/pred,
  fSum.Map (which double-counted coordinate 0) and AStringFunction.TO_CAMEL
  (which read past the end on a trailing underscore).
- Visitor1/Visitor2 called invite(this) back from visit(), and
  Context.send() called itself: unbounded recursion in both.
- Delegate.raiseEvent wrapped the whole dispatch loop in one try/catch, so
  a single failing target silently stopped the chain. The catch is now per
  target and the first failure is rethrown after all targets have run.
- Maths: Logarithm.getDerivative and getFuncDerive disagreed with Map();
  Projections.Mercator indexed V[2] of a 2-element vector; OdeLorentz had a
  sign error against the stated equation; ByRefInt/ByRefLong ROR shifted by
  one bit too many.
- FilterTestWaiter called wait/notify outside a synchronized block.
- MultiCaster.putA started an unbounded thread per item; it now uses a
  shared bounded pool with CallerRunsPolicy for backpressure.
- RandomGUID swallowed a NoSuchAlgorithmException and continued with a
  null digest; it now rethrows.

Two changes go slightly beyond a local fix and want review: Context.send()
required adding send(Context) to the structure.State interface (no
implementors exist in the tree, but it is an interface change), and
Product.getHMV was ambiguous between a wrong name and wrong maths - the
maths is now a real harmonic mean and the previous geometric-mean formula
is kept reachable as getGMV() rather than dropped.

All 39 touched files compile under javac 11 and keep CRLF endings.

Note on verification: javac must be given -encoding UTF-8 for this tree.
Under the platform default (Cp1252 on this machine) files containing
non-ASCII bytes, such as function/string/AStringFunction.java, report
spurious 'illegal character' errors that are an artefact of the decoding,
not defects. This holds at HEAD as well.

+++ technology/xml/XmlUnmarshaller.java
@@ -299,11 +299,7 @@ public class XmlUnmarshaller extends SaxHandler {
 		} else if (argType == int  .class){ newValue = Integer.decode(strValue);
 		} else if (argType == long .class){ newValue = Long.decode(strValue);
 		} else if (argType == float.class){ newValue = Float.valueOf(strValue);
-		// TODO: LOGIC: duplicate check of "argType == long.class" (already handled two branches
-		// above); almost certainly meant "argType == double.class". As written, a double-typed
-		// field is never converted here and falls through to the no-op else branch, leaving
-		// newValue as the raw String instead of a Double.
-		} else if (argType == long .class){ newValue = Double.valueOf(strValue);
+		} else if (argType == double.class){ newValue = Double.valueOf(strValue);
 		} else if (argType == Date .class){ newValue = new Date(Date.parse(newValue.toString()));
 		} else {
 		}
